docs(engine): replace car sample for voxels sample
