+++ engine/samples/CMakeLists.txt
@@ -46,4 +46,4 @@ make_sample(DIR "assets/saving" ASSETS)
 make_sample(DIR "renderer")
 make_sample(DIR "collisions" COMPONENTS)
 make_sample(DIR "scene" COMPONENTS ASSETS)
-make_sample(DIR "cars" COMPONENTS ASSETS)
+make_sample(DIR "voxels" COMPONENTS ASSETS)
