+++ engine/samples/voxels/assets/car.grd
