+++ engine/samples/voxels/assets/car.qb
