+++ engine/samples/voxels/assets/main.pal.meta
